zenn-multi-agents/pyproject.toml
[project]
name = "zenn-multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.2.0",
]
